requirements.txt
pillow
piexif